- Added security features like authentication and authorization using JWTs - Added a user model and new endpoints for user resgistration, login and logout - Changed IDs type to int for query params and secured endpoints access - Included additional claims to JWT - Included endpoints for token refresh

+++ resources/item.py
@@ -1,8 +1,7 @@
-import uuid
-
 from typing import Any
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -12,21 +11,27 @@ from schemas import ItemSchema, ItemUpdateSchema
 item_api = Blueprint("items", __name__, description="Operations on items.")
 
 
-@item_api.route("/item/<string:item_id>")
+@item_api.route("/item/<int:item_id>")
 class Store(MethodView):
+    @jwt_required()
     @item_api.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
 
         return item
 
+    @jwt_required(fresh=True)
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get(item_id)
         db.session.delete(item)
         db.session.commit()
 
         return {"message": "Item deleted."}
 
+    @jwt_required()
     @item_api.arguments(ItemUpdateSchema)
     @item_api.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -47,6 +52,7 @@ class Store(MethodView):
 
 @item_api.route("/item")
 class Store(MethodView):
+    @jwt_required()
     @item_api.arguments(ItemSchema)
     @item_api.response(201, ItemSchema)
     def post(self, item_data):
@@ -60,6 +66,7 @@ class Store(MethodView):
 
         return item
 
+    @jwt_required()
     @item_api.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
